Server Client Files.
These are the original files that I plan on changing. Currently to get this code to work you need to compile all of the files, then run ServerTest and then run ClientTest and you should have two 'chat' boxes. Need to edit code but its a start. If you run into issues with either the server or client not working it might be because the port to which you're connecting from might not be available...Need to look into this as well when adding in the other functionality.